fix: dependencies (#46)
