+++ src/features/shared/index.ts
@@ -1,3 +1,1 @@
-export * from './shared.api';
-export * from './shared.dto';
 export * from './shared.hook';
